+++ packages/broker/test/sequential/plugins/storage/DeleteExpiredCmd.test.ts
@@ -67,7 +67,6 @@ describe('DeleteExpiredCmd', () => {
             auth: {
                 privateKey: mockUser.privateKey
             },
-            restUrl: `http://${STREAMR_DOCKER_DEV_HOST}/api/v2`,
             orderMessages: false,
         })
         deleteExpiredCmd = new DeleteExpiredCmd({