+++ packages/client/test/unit/utils.test.ts
@@ -16,31 +16,18 @@ interface TestResponse {
 
 describeRepeats('utils', () => {
     describe('with mock server', () => {
-        let session: any
         let expressApp: Application
         let server: Server
         const baseUrl = 'http://127.0.0.1:30000'
         const testUrl = '/some-test-url'
 
         beforeEach((done) => {
-            session = {
-                options: {},
-                isUnauthenticated() {
-                    return false
-                },
-                getSessionToken: () => {}
-            }
-
             expressApp = express()
 
             function handle(req: any, res: any) {
-                if (req.get('Authorization') !== 'Bearer session-token') {
-                    res.sendStatus(401)
-                } else {
-                    res.status(200).send({
-                        test: 'test',
-                    })
-                }
+                res.status(200).send({
+                    test: 'test',
+                })
             }
 
             expressApp.get(testUrl, (req: any, res: any) => handle(req, res))
@@ -68,36 +55,8 @@ describeRepeats('utils', () => {
         })
 
         describe('authFetch', () => {
-            it('should return normally when valid session token is passed', async () => {
-                session.getSessionToken = jest.fn(async () => 'session-token')
-                const res = await authFetch<TestResponse>(baseUrl + testUrl, { session })
-                expect(session.getSessionToken).toHaveBeenCalledTimes(1)
-                expect(res.test).toBeTruthy()
-            })
-
-            it('should return 401 error when invalid session token is passed twice', async () => {
-                session.getSessionToken = jest.fn(async () => 'invalid-token')
-                const onCaught = jest.fn()
-                const err = await authFetch<TestResponse>(baseUrl + testUrl, { session }).catch((error) => {
-                    onCaught()
-                    return error
-                })
-
-                expect(session.getSessionToken).toHaveBeenCalledTimes(2)
-                expect(err.toString()).toMatch(
-                    '401. \'Unauthorized\'',
-                )
-                expect(err.body).toEqual('Unauthorized')
-                expect(onCaught).toHaveBeenCalledTimes(1)
-            })
-
-            it('should return normally when valid session token is passed after expired session token', async () => {
-                session.getSessionToken = jest.fn()
-                    .mockImplementationOnce(async () => 'expired-session-token')
-                    .mockImplementationOnce(async () => 'session-token')
-
-                const res = await authFetch<TestResponse>(baseUrl + testUrl, { session })
-                expect(session.getSessionToken).toHaveBeenCalledTimes(2)
+            it('should return normally', async () => {
+                const res = await authFetch<TestResponse>(baseUrl + testUrl)
                 expect(res.test).toBeTruthy()
             })
         })